Formatting pass

+++ source_files/obsidian_main/g_nukem.cc
@@ -21,13 +21,12 @@
 
 #include "g_nukem.h"
 
-#include "headers.h"
 #include <algorithm>
 
 #include "csg_main.h"
 #include "hdr_lua.h"
+#include "headers.h"
 #include "images.h"
-
 #include "lib_grp.h"
 #include "lib_util.h"
 #include "m_cookie.h"
@@ -35,7 +34,7 @@
 #include "main.h"
 #include "q_common.h"  // qLump_c
 
-extern void CSG_NUKEM_Write();
+extern void                  CSG_NUKEM_Write();
 extern std::filesystem::path BestDirectory();
 
 // Properties
@@ -47,7 +46,8 @@ static qLump_c *nk_sprites;
 
 static raw_nukem_map_t nk_header;
 
-static void NK_FreeLumps() {
+static void NK_FreeLumps()
+{
     delete nk_sectors;
     nk_sectors = nullptr;
     delete nk_walls;
@@ -56,15 +56,18 @@ static void NK_FreeLumps() {
     nk_sprites = nullptr;
 }
 
-static void NK_WriteLump(const char *name, qLump_c *lump) {
+static void NK_WriteLump(const char *name, qLump_c *lump)
+{
     SYS_ASSERT(strlen(name) <= 11);
 
     GRP_NewLump(name);
 
     int len = lump->GetSize();
 
-    if (len > 0) {
-        if (!GRP_AppendData(lump->GetBuffer(), len)) {
+    if (len > 0)
+    {
+        if (!GRP_AppendData(lump->GetBuffer(), len))
+        {
             //    errors_seen++;
         }
     }
@@ -72,10 +75,9 @@ static void NK_WriteLump(const char *name, qLump_c *lump) {
     GRP_FinishLump();
 }
 
-bool NK_StartGRP(const std::filesystem::path &filename) {
-    if (!GRP_OpenWrite(filename)) {
-        return false;
-    }
+bool NK_StartGRP(const std::filesystem::path &filename)
+{
+    if (!GRP_OpenWrite(filename)) { return false; }
 
     qLump_c *info = BSP_CreateInfoLump();
     NK_WriteLump("OBLIGE.DAT", info);
@@ -86,9 +88,10 @@ bool NK_StartGRP(const std::filesystem::path &filename) {
 
 void NK_EndGRP(void) { GRP_CloseWrite(); }
 
-void NK_BeginLevel(const char *level_name) {
+void NK_BeginLevel(const char *level_name)
+{
     std::string lump_name = StringFormat("%s.MAP", level_name);
-    lump_name = StringUpper(lump_name.c_str());
+    lump_name             = StringUpper(lump_name.c_str());
 
     GRP_NewLump(lump_name.c_str());
 
@@ -101,15 +104,16 @@ void NK_BeginLevel(const char *level_name) {
     NK_FreeLumps();
 
     nk_sectors = new qLump_c;
-    nk_walls = new qLump_c;
+    nk_walls   = new qLump_c;
     nk_sprites = new qLump_c;
 }
 
-void NK_EndLevel() {
+void NK_EndLevel()
+{
     // write everything...
 
     uint16_t num_sectors = LE_U16(NK_NumSectors());
-    uint16_t num_walls = LE_U16(NK_NumWalls());
+    uint16_t num_walls   = LE_U16(NK_NumWalls());
     uint16_t num_sprites = LE_U16(NK_NumSprites());
 
     GRP_AppendData(&nk_header, (int)sizeof(nk_header));
@@ -130,7 +134,8 @@ void NK_EndLevel() {
 
 void NK_AddSector(int first_wall, int num_wall, int visibility, int f_h,
                   int f_pic, int c_h, int c_pic, int c_flags, int lo_tag,
-                  int hi_tag) {
+                  int hi_tag)
+{
     raw_nukem_sector_t raw;
 
     memset(&raw, 0, sizeof(raw));
@@ -139,30 +144,29 @@ void NK_AddSector(int first_wall, int num_wall, int visibility, int f_h,
     raw.wall_num = LE_U16(num_wall);
 
     raw.floor_h = LE_S32(f_h);
-    raw.ceil_h = LE_S32(c_h);
+    raw.ceil_h  = LE_S32(c_h);
 
     raw.floor_pic = LE_U16(f_pic);
-    raw.ceil_pic = LE_U16(c_pic);
+    raw.ceil_pic  = LE_U16(c_pic);
 
     raw.ceil_flags = LE_U16(c_flags);
 
     // preven the space skies from killing the player
-    if (c_flags & SECTOR_F_PARALLAX) {
-        raw.ceil_palette = 3;
-    }
+    if (c_flags & SECTOR_F_PARALLAX) { raw.ceil_palette = 3; }
 
     raw.visibility = visibility;
 
     raw.lo_tag = LE_U16(lo_tag);
     raw.hi_tag = LE_U16(hi_tag);
-    raw.extra = LE_U16(-1);
+    raw.extra  = LE_U16(-1);
 
     nk_sectors->Append(&raw, sizeof(raw));
 }
 
 void NK_AddWall(int x, int y, int right, int back, int back_sec, int flags,
                 int pic, int mask_pic, int xscale, int yscale, int xpan,
-                int ypan, int lo_tag, int hi_tag) {
+                int ypan, int lo_tag, int hi_tag)
+{
     raw_nukem_wall_t raw;
 
     memset(&raw, 0, sizeof(raw));
@@ -173,26 +177,27 @@ void NK_AddWall(int x, int y, int right, int back, int back_sec, int flags,
     raw.right_wall = LE_U16(right);
 
     raw.back_wall = LE_U16(back);
-    raw.back_sec = LE_U16(back_sec);
+    raw.back_sec  = LE_U16(back_sec);
 
-    raw.pic = LE_U16(pic);
+    raw.pic      = LE_U16(pic);
     raw.mask_pic = LE_U16(mask_pic);
-    raw.flags = LE_U16(flags);
+    raw.flags    = LE_U16(flags);
 
     raw.xscale = xscale;
-    raw.xpan = xpan;
+    raw.xpan   = xpan;
     raw.yscale = yscale;
-    raw.ypan = ypan;
+    raw.ypan   = ypan;
 
     raw.lo_tag = LE_U16(lo_tag);
     raw.hi_tag = LE_U16(hi_tag);
-    raw.extra = LE_U16(-1);
+    raw.extra  = LE_U16(-1);
 
     nk_walls->Append(&raw, sizeof(raw));
 }
 
 void NK_AddSprite(int x, int y, int z, int sec, int flags, int pic, int angle,
-                  int lo_tag, int hi_tag) {
+                  int lo_tag, int hi_tag)
+{
     raw_nukem_sprite_t raw;
 
     memset(&raw, 0, sizeof(raw));
@@ -201,39 +206,42 @@ void NK_AddSprite(int x, int y, int z, int sec, int flags, int pic, int angle,
     raw.y = LE_S32(y);
     raw.z = LE_S32(z);
 
-    raw.flags = LE_U16(flags);
-    raw.pic = LE_U16(pic);
-    raw.angle = LE_U16(angle);
+    raw.flags  = LE_U16(flags);
+    raw.pic    = LE_U16(pic);
+    raw.angle  = LE_U16(angle);
     raw.sector = LE_U16(sec);
 
-    raw.xscale = 40;
-    raw.yscale = 40;
+    raw.xscale    = 40;
+    raw.yscale    = 40;
     raw.clip_dist = 32;
 
     raw.lo_tag = LE_U16(lo_tag);
     raw.hi_tag = LE_U16(hi_tag);
-    raw.extra = LE_U16(-1);
+    raw.extra  = LE_U16(-1);
 
     raw.owner = -1;
 
-    if (NK_NumSprites() == 0 || pic == 1405 /* APLAYER */) {
-        nk_header.pos_x = raw.x;
-        nk_header.pos_y = raw.y;
-        nk_header.pos_z = raw.z;
-        nk_header.angle = raw.angle;
+    if (NK_NumSprites() == 0 || pic == 1405 /* APLAYER */)
+    {
+        nk_header.pos_x  = raw.x;
+        nk_header.pos_y  = raw.y;
+        nk_header.pos_z  = raw.z;
+        nk_header.angle  = raw.angle;
         nk_header.sector = raw.sector;
     }
 
     nk_sprites->Append(&raw, sizeof(raw));
 }
 
-int NK_NumSectors() {
+int NK_NumSectors()
+{
     return nk_sectors->GetSize() / sizeof(raw_nukem_sector_t);
 }
 
 int NK_NumWalls() { return nk_walls->GetSize() / sizeof(raw_nukem_wall_t); }
 
-int NK_NumSprites() {
+int NK_NumSprites()
+{
     return nk_sprites->GetSize() / sizeof(raw_nukem_sprite_t);
 }
 
@@ -241,7 +249,8 @@ int NK_NumSprites() {
 //  GAME  INTERFACE
 //------------------------------------------------------------------------
 
-class nukem_game_interface_c : public game_interface_c {
+class nukem_game_interface_c : public game_interface_c
+{
    private:
     std::filesystem::path filename;
 
@@ -251,44 +260,54 @@ class nukem_game_interface_c : public game_interface_c {
     bool Start(const char *preset);
     bool Finish(bool build_ok);
 
-    void BeginLevel();
-    void EndLevel();
-    void Property(std::string key, std::string value);
+    void                  BeginLevel();
+    void                  EndLevel();
+    void                  Property(std::string key, std::string value);
     std::filesystem::path Filename();
     std::filesystem::path ZIP_Filename();
 
    private:
 };
 
-bool nukem_game_interface_c::Start(const char *preset) {
+bool nukem_game_interface_c::Start(const char *preset)
+{
 #ifndef CONSOLE_ONLY
-    if (batch_mode) {
-        if (batch_output_file.is_absolute()) {
-            filename = batch_output_file;
-        } else {
-            filename = std::filesystem::current_path().append(batch_output_file.string());
+    if (batch_mode)
+    {
+        if (batch_output_file.is_absolute()) { filename = batch_output_file; }
+        else
+        {
+            filename = std::filesystem::current_path().append(
+                batch_output_file.string());
         }
-    } else {
-        filename = std::filesystem::current_path().append(preset).replace_extension("grp").u8string().c_str();
+    }
+    else
+    {
+        filename = std::filesystem::current_path()
+                       .append(preset)
+                       .replace_extension("grp")
+                       .u8string()
+                       .c_str();
     }
 #else
-    if (batch_output_file.is_absolute()) {
-        filename = batch_output_file;
-    } else {
-        filename = std::filesystem::current_path().append(batch_output_file.string());
+    if (batch_output_file.is_absolute()) { filename = batch_output_file; }
+    else
+    {
+        filename =
+            std::filesystem::current_path().append(batch_output_file.string());
     }
 #endif
 
-    if (filename.empty()) {
+    if (filename.empty())
+    {
         Main::ProgStatus(_("Cancelled"));
         return false;
     }
 
-    if (create_backups) {
-        Main::BackupFile(filename);
-    }
+    if (create_backups) { Main::BackupFile(filename); }
 
-    if (!NK_StartGRP(filename.c_str())) {
+    if (!NK_StartGRP(filename.c_str()))
+    {
         Main::ProgStatus(_("Error (create file)"));
         return false;
     }
@@ -296,42 +315,42 @@ bool nukem_game_interface_c::Start(const char *preset) {
     return true;
 }
 
-bool nukem_game_interface_c::Finish(bool build_ok) {
+bool nukem_game_interface_c::Finish(bool build_ok)
+{
     NK_EndGRP();
 
     // remove the file if an error occurred
-    if (!build_ok) {
-        std::filesystem::remove(filename);
-    } else {
-        Recent_AddFile(RECG_Output, filename);
-    }
+    if (!build_ok) { std::filesystem::remove(filename); }
+    else { Recent_AddFile(RECG_Output, filename); }
 
     return build_ok;
 }
 
 void nukem_game_interface_c::BeginLevel() {}
 
-void nukem_game_interface_c::Property(std::string key, std::string value) {
-    if (StringCaseCmp(key, "level_name") == 0) {
-        level_name = value.c_str();
-    } else if (StringCaseCmp(key, "description") == 0) {
+void nukem_game_interface_c::Property(std::string key, std::string value)
+{
+    if (StringCaseCmp(key, "level_name") == 0) { level_name = value.c_str(); }
+    else if (StringCaseCmp(key, "description") == 0)
+    {
         // ignored (for now)
         // [another mechanism sets the description via BEX/DDF]
-    } else {
-        LogPrintf("WARNING: unknown NUKEM property: %s=%s\n", key.c_str(), value.c_str());
+    }
+    else
+    {
+        LogPrintf("WARNING: unknown NUKEM property: %s=%s\n", key.c_str(),
+                  value.c_str());
     }
 }
 
-std::filesystem::path nukem_game_interface_c::Filename() {
-    return filename;
-}
+std::filesystem::path nukem_game_interface_c::Filename() { return filename; }
 
-std::filesystem::path nukem_game_interface_c::ZIP_Filename() {
-    return "";
-}
+std::filesystem::path nukem_game_interface_c::ZIP_Filename() { return ""; }
 
-void nukem_game_interface_c::EndLevel() {
-    if (level_name.empty()) {
+void nukem_game_interface_c::EndLevel()
+{
+    if (level_name.empty())
+    {
         Main::FatalError("Script problem: did not set level name!\n");
     }
